Extend source files to include property-lists/strings
This allows .plists and .strings to be searched for unused resources.

This inclusion, however, also causes the often-seen "Info.plist" to appear as a dangling file.

This is a special case, and it can might warrant a specific exclusion rule for this particular file. Alternatively, search the project.pbxproj for `INFOPLIST_FILE` and treat that as target membership. This is similar to the "AppIcon" asset issue.