Add edge-case in resolving file URLs
This commit will treat any group with `name` set  as a non-folder group, ignoring the path component.

This might not be right, but I did find a .pbxproj where a group had both `path` and `name` set, and in this case the `path` was very obviously not correct, but caused no issues in Xcode.