Revert edge-case in resolving file URLs
I found another project with a similar setup, but with completely valid `name` and `path`. I think the correct way is to just assume that `path` *is* actually on disk, and if not, the group is invalid (e.g. another defect, which, if fixed, cascades to all the files with incorrectly mapped paths).