Refactor and add progress indication for emptyAssets
Progress indication adds a lot of code but provides little value for many of the defects we check for.

Generally, they are only useful for the potentially time-consuming checks (e.g. those involving opening and reading files). For the quicker checks, you will typically only ever see the total anyway, so prefer only implementing progress indication for a select few checks where it makes sense.

If we _really_ wanted to know the count, we could just add the total at the end of a check. However, the count is also not very useful; what does it count? For example, `unusedResources` does not count the number of resources, but rather the number of source files to go through.

Closes #3