Add storyboard/xib as resources
A note on xib's: these are tricky to find usages of.

It's common practice to match a view up with a xib, naming both identically; for example, `MyView: UIView` and `MyView.xib`. It is also common to then not load this xib by string (i.e. `loadNib("MyView")`, but rather `MyView.className` (or similar implementation).
This case would not be discovered by the (intentionally) simple quoted search we have going on.

However, since we _are_ also searching the xib file itself, we will find an attribute `customClass="MyView"` which _does_ match the search. This is obviously not a guarantee that the class is ever instantiated (out of scope for this project), but if we assume the aforementioned convention, it is a good enough indicator that the resource is in use.

The search is already inherently error-prone, so I prefer the simplicity of accepting this solution over trying to configure elaborate and complicated searches.